Big refactor to naming and package convections.
